requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
idna==2.8
lxml==4.4.2
Pillow==7.0.0
python-docx==0.8.10
requests==2.22.0
soupsieve==1.9.5
urllib3==1.25.7
EZGmail==0.0.8
